pyproject.toml
[tool.poetry]
name = "Search User Gists"
version = "0.1.0"
description = "Searching User Github Gists"
authors = ["Jatin"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "2.3.2"
Werkzeug = "2.3.6"
requests = "2.28.1"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-mock = "^3.0.0"
flake8 = "^6.0.0"
black = "^23.0.0"
isort = "^5.0.0"
pre-commit = "^3.0.0"
mypy = "^1.0.0"
mypy-extensions = "^1.0.0"
autopep8 = "^1.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88

[tool.mypy]
python_version = "3.9"
strict = true
